requirements.txt
newspaper3k
tfds-nightly
torch
transformers
nltk
textblob